+++ Game/Networking/Protocol.hpp
@@ -16,7 +16,7 @@ namespace net
 	{
 	public:
 
-		AbstractProtocol(unsigned int, const AbstractExportStrategy&, const AbstractImportStrategy&);
+		AbstractProtocol(unsigned short, const AbstractExportStrategy&, const AbstractImportStrategy&);
 		virtual ~AbstractProtocol();
 		virtual void Tick(def::time) = 0; //Handle incoming packets, forward them via the specified exportstrategy.
 		virtual void Broadcast() = 0; //Import packets via the specified importstrategy, broadcast them to the peers (includes congestion avoidance).
@@ -36,7 +36,7 @@ namespace net
 
 	public:
 
-		ClientsideProtocol(unsigned int, const AbstractExportStrategy&, const AbstractImportStrategy&);
+		ClientsideProtocol(unsigned short, const AbstractExportStrategy&, const AbstractImportStrategy&);
 		ClientsideProtocol(const AbstractExportStrategy&, const AbstractImportStrategy&);
 		ClientsideProtocol() = delete;
 		virtual ~ClientsideProtocol() override;
@@ -54,7 +54,7 @@ namespace net
 	{
 	public:
 
-		ServersideProtocol(unsigned int, const AbstractExportStrategy&, const AbstractImportStrategy&);
+		ServersideProtocol(unsigned short, const AbstractExportStrategy&, const AbstractImportStrategy&);
 		ServersideProtocol(const AbstractExportStrategy&, const AbstractImportStrategy&);
 		ServersideProtocol() = delete;
 		virtual ~ServersideProtocol() override;
